requirements.txt
click
requests
rich
PyYAML
tabulate